# Memory optimization for binned analyses (for combination)