--- a/interface/CMSHistFunc.h
+++ b/interface/CMSHistFunc.h
@@ -146,6 +146,7 @@ class CMSHistFunc : public RooAbsReal {
 
   static void EnableFastVertical();
   friend class CMSHistV<CMSHistFunc>;
+  friend class CMSHistSum;
 
   /*