--- a/interface/CMSHistSum.h
+++ b/interface/CMSHistSum.h
@@ -34,6 +34,7 @@ private:
     std::vector<RooRealVar*> push_res;
   };
 public:
+
   CMSHistSum();
 
   CMSHistSum(const char* name, const char* title, RooRealVar& x,
@@ -46,8 +47,6 @@ public:
   }
 
   virtual ~CMSHistSum() {;}
-
-  void applyErrorShifts(unsigned idx, FastHisto const& nominal, FastHisto & result);
 
   Double_t evaluate() const;
 
@@ -69,45 +68,76 @@ public:
 
   inline FastHisto const& cache() const { return cache_; }
 
-  RooArgList wrapperList() const;
-  RooArgList const& coefList() const { return coeffs_; }
-  RooArgList const& funcList() const { return funcs_; }
+  RooArgList const& coefList() const { return coeffpars_; }
+  // RooArgList const& funcList() const { return funcs_; }
 
+  static void EnableFastVertical();
   friend class CMSHistV<CMSHistSum>;
 
  protected:
   RooRealProxy x_;
-  RooListProxy funcs_;
-  RooListProxy coeffs_;
+
+  RooListProxy morphpars_;
+  RooListProxy coeffpars_;
   RooListProxy binpars_;
-  mutable std::vector<CMSHistFunc const*> vfuncs_; //!
-  mutable std::vector<RooAbsReal const*> vcoeffs_; //!
+
+  int n_procs_;
+  int n_morphs_;
+
+  std::vector<FastTemplate> storage_;  // All nominal and vmorph templates
+  std::vector<int> process_fields_; // Indicies for process templates in storage_
+  std::vector<int> vmorph_fields_; // Indicies for vmorph templates in storage_
+
+  std::vector<FastTemplate> binerrors_; // Bin errors for each process
+
+  std::vector<CMSHistFunc::VerticalSetting> vtype_; // Vertical morphing type for each process
+  std::vector<double> vsmooth_par_; // Vertical morphing smooth region for each process
+
+  mutable std::vector<CMSHistFunc const*> vfuncstmp_; //!
+  mutable std::vector<RooAbsReal const*> vcoeffpars_; //!
+  mutable std::vector<RooAbsReal const*> vmorphpars_; //!
   mutable std::vector<std::vector<RooAbsReal *>> vbinpars_; //!
   std::vector<std::vector<unsigned>> bintypes_;
 
   mutable std::vector<double> coeffvals_; //!
+
+  mutable std::vector<FastHisto> compcache_; //!
+  mutable FastHisto staging_; //!
   mutable FastHisto valsum_; //!
-  mutable FastHisto cache_; //!
+  mutable FastHisto cache_;
+
   mutable std::vector<double> err2sum_; //!
   mutable std::vector<double> toterr_; //!
   mutable std::vector<std::vector<double>> binmods_; //!
   mutable std::vector<std::vector<double>> scaledbinmods_; //!
+
   mutable SimpleCacheSentry sentry_; //!
   mutable SimpleCacheSentry binsentry_; //!
+
   mutable std::vector<double> data_; //!
 
   mutable BarlowBeeston bb_; //!
 
   mutable bool initialized_; //! not to be serialized
 
-  mutable int last_eval_; //! not to be serialized
-
   mutable bool analytic_bb_; //! not to be serialized
 
+  mutable std::vector<double> vertical_prev_vals_; //! not to be serialized
+  mutable int fast_mode_; //! not to be serialized
+  static bool enable_fast_vertical_; //! not to be serialized
+
+  inline int& morphField(int const& ip, int const& iv) {
+    return vmorph_fields_[ip * n_morphs_ + iv];
+  }
+
   void initialize() const;
-  void updateCache(int eval = 1) const;
+  void updateCache() const;
+  inline double smoothStepFunc(double x, int const& ip) const;
+
 
   void runBarlowBeeston() const;
+
+  void updateMorphs() const;
 
 
  private: