Adds new CMSHistSum class. Essentially a clone of CMSHistErrorPropagator, but instead of linking to a set of CMSHistFuncs (one per process), encapsulates all processes and performs the vertical template morphing directly. For complex models (e.g. STXS) this dramatically reduces the number of separate RooFit objects that need to go into the workspace. Should be used with `--X-rtd FAST_VERTICAL_MORPH`, as otherwise fitting will be slightly slower than traditional approach, as we now have a single sentry watching all vertical morph params for a given channel.

In testing on the full htautau cards, reduces fit memory usage by over 60%, and fitting time by 40%. Unfortunately not possible to improve t2w memory at this point. We still build full set of CMSHistFuncs, then CMSHistSum constructor extracts all the needed info. A more optimal way would require extensive rewrite of t2w.

Also confirmed that for performing simple fits, there is no need for CMSHistSum to be cloned inside of CachingPdf. text2workspace.py will now add an attribute on CMSHistSum, that in conjunction with `--X-rtd CACHINGPDF_NOCLONE` will prevent the clone.

Recommend validating carefully on all other channels before general use. Results should be highly consistent (at least 3 digits).

**Testing on the full htautau cards: ML_STXS/stage1/125/**

`combineCards.py htt*.txt > combined_all.txt`

OLD
------
`text2workspace.py -o combined_all.root -m 125 --for-fits --no-wrappers --X-pack-asympows --optimize-simpdf-constraints=cms combined_all.txt`
=> 2619.49s / 6.2 GB

`combine -M MultiDimFit --algo none --noMCbonly 1 --cminDefaultMinimizerStrategy 0 --X-rtd NO_INITIAL_SNAP --X-rtd OPTIMIZE_BOUNDS=0 combined_all.root -v 3 --X-rtd FAST_VERTICAL_MORPH`
=> 962.54s / 2.3 GB
```
Minimized in : Real time 0:14:52, CP time 891.940
FINAL NLL - NLL0 VALUE = -1562.217683
 --- MultiDimFit ---
best fit parameter values:
   r :    +0.912
Done in 15.71 min (cpu), 15.71 min (real)
```

NEW
-------
`text2workspace.py -o combined_all_new.root -m 125 --for-fits --no-wrappers --X-pack-asympows --optimize-simpdf-constraints=cms combined_all.txt --use-histsum`
=> 2358.91s / 6.1 GB

`combine -M MultiDimFit --algo none --noMCbonly 1  --cminDefaultMinimizerStrategy 0 --X-rtd NO_INITIAL_SNAP --X-rtd OPTIMIZE_BOUNDS=0 combined_all_new.root -v 3 --X-rtd FAST_VERTICAL_MORPH --X-rtd CACHINGPDF_NOCLONE`
=> 566s / 0.87 GB
```
Minimized in : Real time 0:08:58, CP time 537.780
FINAL NLL - NLL0 VALUE = -1562.217603

 --- MultiDimFit ---
best fit parameter values:
   r :    +0.912
Done in 9.26 min (cpu), 9.27 min (real)
```