@adewit 2021-08-23
I've tested this on the binned STXS channels (just the best fits). Summary: generally no or very small differences observed, with differences in the 3d digit for ttHbb which were unexpected. In constructing the bbb parameters there's a message
```
  ----------------------------------------------------------
    ttbarW_CR            -0.018423       0.016337
      => Cannot handle negative content, ignore
  ----------------------------------------------------------
```
 With the new CMSHistFunc class, whereas with the old class we had
```
  ----------------------------------------------------------
    ttbarW_CR            0.000000        0.016337
      => Poisson not viable, prop_binSTXS_2017_fh_STXS_3_j7_t4_DNN_Node0_bin9_ttbarW_CR[0.00,-7.00,7.00] to be gaussian constrained
  ----------------------------------------------------------
```
So there's a difference in the underflow cropping at this stage. 

There are differences in the 4th digit for the H->tt STXS fit (it's not exactly the same issue as for ttHbb, ie the number of parameters in the fit is the same in both cases here.)

Results of the comparison (difference between best fit values using old CMSHistFunc and using new CMSHistFunc) below. 

For vhbb:
![vhbb_diffresults_testmods](https://user-images.githubusercontent.com/5803497/130406446-e12b9d6b-bb3f-47c0-b814-a283190a9212.png)

hww:
![hww_diffresults_testmods](https://user-images.githubusercontent.com/5803497/130406449-d1daa0ff-dcbd-4363-92d5-66b79563b36b.png)

htt:
![htt_stxs_diffresults_testmods](https://user-images.githubusercontent.com/5803497/130406451-2508dc7d-efb5-4fe8-997e-db6ad7b80d66.png)

tthbb:
![tth_hbb_diffresults_testmods](https://user-images.githubusercontent.com/5803497/130406442-fb046cad-2ea1-48e3-8eed-6e6ddf46e258.png)