@adewit 2021-08-23
Yes, tth bb looks consistent now:
![tth_hbb_diffresults_testmods](https://user-images.githubusercontent.com/5803497/130438350-0dfeae83-cba6-4e8a-983e-7f94e08da287.png)

Just for the record, this doesn't change the small differences in the H->tt results, but since those differences are < 0.5 per mille, I think we can live with that :-) :
![htt_stxs_diffresults_testmods](https://user-images.githubusercontent.com/5803497/130438356-de469bfd-c671-4e6e-81f0-572b4ed385ff.png)